+++ pyllars/mpl_utils.py
@@ -980,6 +980,8 @@ def plot_confusion_matrix(
         show_colorbar:bool=True,
         title:Optional[str]="Confusion matrix", 
         cmap:matplotlib.colors.Colormap=plt.cm.Blues, 
+        vmin:Optional[float]=None,
+        vmax:Optional[float]=None,
         true_tick_labels:Optional[Sequence[str]]=None, 
         predicted_tick_labels:Optional[Sequence[str]]=None, 
         ylabel:Optional[str]="True labels", 
@@ -1046,7 +1048,7 @@ def plot_confusion_matrix(
     if cmap == None:
         cmap = plt.cm.Blues
 
-    mappable = ax.imshow(confusion_matrix, interpolation='nearest', cmap=cmap)
+    mappable = ax.imshow(confusion_matrix, interpolation='nearest', cmap=cmap, vmin=vmin, vmax=vmax)
     
     if show_colorbar:
         fig.colorbar(mappable)
@@ -1241,8 +1243,8 @@ def plot_roc_curve(
         line_colors:Optional[Sequence]=None,
         point_colors:Optional[Sequence]=None,
         alphas:Optional[Sequence[float]]=None,
-        line_kwargs:Optional[Mapping]=None,
-        point_kwargs:Optional[Mapping]=None,
+        line_kwargs:Optional[Mapping]={},
+        point_kwargs:Optional[Mapping]={},
         title:Optional[str]="Receiver operating characteristic curves",
         xlabel:Optional[str]="False positive rate",
         ylabel:Optional[str]="True positive rate",